+++ eds-user-20210308/alibabacloud_eds_user20210308/models/_get_users_count_request.py
@@ -1,0 +1,49 @@
+# -*- coding: utf-8 -*-
+# This file is auto-generated, don't edit it. Thanks.
+from __future__ import annotations
+
+from darabonba.model import DaraModel
+
+class GetUsersCountRequest(DaraModel):
+    def __init__(
+        self,
+        biz_type: str = None,
+        business_channel: str = None,
+        solution_id: str = None,
+    ):
+        self.biz_type = biz_type
+        self.business_channel = business_channel
+        self.solution_id = solution_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.biz_type is not None:
+            result['BizType'] = self.biz_type
+
+        if self.business_channel is not None:
+            result['BusinessChannel'] = self.business_channel
+
+        if self.solution_id is not None:
+            result['SolutionId'] = self.solution_id
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('BizType') is not None:
+            self.biz_type = m.get('BizType')
+
+        if m.get('BusinessChannel') is not None:
+            self.business_channel = m.get('BusinessChannel')
+
+        if m.get('SolutionId') is not None:
+            self.solution_id = m.get('SolutionId')
+
+        return self
+